# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/tasks/{uuid}": {
            "get": {
                "description": "Find one task by UUID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "Find One Task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task",
                        "schema": {
                            "$ref": "#/definitions/handler.TaskResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete task by UUID",
                "tags": [
                    "Task"
                ],
                "summary": "Delete task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.CategoryResponse": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
...
        "handler.TaskResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/handler.CategoryResponse"
                },
                "description": {
                    "type": "string"
                },
                "done": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        }
    }
}